refactor(server): extract LimitedCurlAsyncHTTPClient into http_client module
