+++ web_monitoring_diff/server/server.py
@@ -22,6 +22,7 @@ import traceback
 import web_monitoring_diff
 from .mock_http import MockResponse
 from .errors import PublicError
+from .http_client import LimitedCurlAsyncHTTPClient
 from .. import basic_diffs, html_render_diff, html_links_diff
 from ..exceptions import UndiffableContentError, UndecodableContentError
 from ..utils import shutdown_executor_in_loop, Signal
@@ -101,27 +102,6 @@ try:
 except ValueError:
     print('DIFFER_MAX_BODY_SIZE must be an integer', file=sys.stderr)
     sys.exit(1)
-
-
-class LimitedCurlAsyncHTTPClient(CurlAsyncHTTPClient):
-    """
-    A customized version of Tornado's CurlAsyncHTTPClient that adds support for
-    maximum response body sizes. The API is the same as that for Tornado's
-    SimpleAsyncHTTPClient: set ``max_body_size`` to an integer representing the
-    maximum number of bytes in a response body.
-    """
-    def initialize(self, max_clients=10, defaults=None, max_body_size=None):
-        self.max_body_size = max_body_size
-        defaults = defaults or {}
-        defaults['prepare_curl_callback'] = self.prepare_curl
-        super().initialize(max_clients=max_clients, defaults=defaults)
-
-    def prepare_curl(self, curl):
-        if self.max_body_size:
-            # NOTE: cURL's docs suggest this doesn't work if the server doesn't
-            # send a Content-Length header, but it seems to do just fine in
-            # tests. ¯\_(ツ)_/¯
-            curl.setopt(pycurl.MAXFILESIZE, self.max_body_size)
 
 
 HTTP_CLIENT = LimitedCurlAsyncHTTPClient
